Moved Open Tree Taxonomy methods into App.vue for better organization.
